requirements.txt
tqdm==4.62.3
numpy==1.20.3
Pillow==8.3.2
matplotlib==3.4.3
opencv-python==4.5.3.56
torch==1.9.1+cu111
torchvision==0.10.1+cu111
torchattacks==3.2.2